requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
flask_wtf
wtforms
mysqlclient
python-dotenv